Conflicts:
	src/main/java/com/clevel/selos/controller/BankStatementDetail.java
	src/main/java/com/clevel/selos/integration/ncb/ncrs/service/NCRSImp.java
	src/main/java/com/clevel/selos/transform/business/ExistingCreditTransform.java

+++ src/main/java/com/clevel/selos/dao/working/DBRDetailDAO.java
@@ -6,7 +6,7 @@ import org.slf4j.Logger;
 
 import javax.inject.Inject;
 
-public class DBRDetailDAO extends GenericDAO<DBRDetail, Integer> {
+public class DBRDetailDAO extends GenericDAO<DBRDetail, Long> {
     @Inject
     private Logger log;
